+++ R/misc.R
@@ -33,3 +33,25 @@ reduce_multiplication <- function(x) {
 reduce_and <- function(x) {
   reduce(x, expr_and)
 }
+
+expr_or <- function(x, y) {
+  expr(!!x | !!y)
+}
+
+reduce_or <- function(x) {
+  reduce(x, expr_or)
+}
+
+combine_path_conditions <- function(conditions) {
+  n <- length(conditions)
+  if (n == 0) {
+    return(TRUE)
+  }
+  if (n == 1) {
+    return(conditions[[1]])
+  }
+  if (n == 2) {
+    return(expr_and(conditions[[1]], conditions[[2]]))
+  }
+  reduce_and(conditions)
+}